Back order Prediction
Approach: The data was obtained from Kaggle. It was treated with null value imputation and feature engineering to make it suitable for model building. After Scaling and Encoding, the data was balanced using SMOTE oversampling. Afterwards, it was modelled using various classification algorithms. Adaboost was chosen as the final model because of its high F1 score.