MOD Clothing EDA_STATS
Let’s assume you are working as a data scientist in a newly started textile company. 
You have given a dataset that consists of most of the features related to the 
measurements, clothes types, Product ID, User ID, etc. Now you suppose to work the 
dataset to identify the patterns to understand the customer’s preferred brands and how 
the company needs to brand its products based on the patterns that you will be found 
them out. Explore all the EDA concepts you learned and use a statistical test to ensure 
that your identification is true or false.